requirements.txt
streamlit
matplotlib
pandas
seaborn
plotly.express
folium